Rename busymachines.effects.result -> busymachines.effects.sync
