+++ effects/src/main/scala/busymachines/effects/optionOpsSyntax.scala
@@ -2,7 +2,7 @@ package busymachines.effects
 
 import busymachines.core.Anomaly
 import busymachines.effects.future.FutureUtil
-import busymachines.effects.result.ResultOps
+import busymachines.effects.sync.ResultOps
 
 import scala.util._
