+++ semver-parsers/src/main/scala/busymachines/semver/SemanticVersionParsers.scala
@@ -1,6 +1,6 @@
 package busymachines.semver
 
-import busymachines.effects.result._
+import busymachines.effects.sync._
 
 /**
   *
